Rename jwk_storage_document_ext module, too
